style: mention ami in script name
I do this because the cuboid model does not represent an AMI setting
(two muscles one tendon). This is why I want to do the bayesian
optimization for cuboid in a seperate script. This will result in
duplicate code. However I think it is still better than rewriting one shared
main script everytime you want to change the model or introduce new
complexity to the
script to handle different sample and bounds shapes.