feat: share cuboid BO simulation results
For every single simulation, the water density parameter rho and Am were
changed (I have no idea what Am does but Georg said it can be seen as a
random effect).

All of the .csv files have some duplicate endings. This is due to the
stopping criterion of the Bayesian Optimization loop still being
configured to the Hill-Type case. Just remove the duplicate samples if
you wish too.

I am unsure how valuable this data is in terms of Data Augmentation
since the optimal pre-stretch is always the maximum allowable force
(10N). In my opinion, this is no surprise and has to do with the
limitations of the simulation setting that only features one muscle with
only one side being attached to a rigid body.

Anyway, I am super happy about this commit and proud of the team that we
were able to actually get Bayesian Optimization for an
opendihu-simulation, the 4x4 Cuboid model, running. THIS IS A BIG ACHIEVEMENT ^^